add pytest tests and tests for datatables api
