+++ league/templates/league/includes/sgrfs_datatable.html
@@ -80,7 +80,6 @@
                 } );
               }
             }
-            //"data": {{ sgfs_data_json | safe }},
         });
     });
 </script>
