+++ pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = openstudyroom.settings.test
+addopts = --reuse-db
