+++ requirements_dev.txt
@@ -1,0 +1,7 @@
+-r requirements.txt
+freezegun==0.3.15
+pytest==5.4.2
+pytest-django==3.9.0
+python-dateutil==2.8.1
+snapshottest==0.5.1
+requests-mock==1.8.0
